+++ types/eslint/rules/best-practices.d.ts
