+++ types/eslint/rules/deprecated.d.ts
