+++ types/eslint/rules/ecmascript-6.d.ts
