+++ types/eslint/rules/index.d.ts
