+++ types/eslint/rules/strict-mode.d.ts
