+++ types/eslint/rules/variables.d.ts
